core/internal/features/ocr2: isolate non-parallel test in separate package
