+++ core/internal/features/ocr2/plugins/plugins.go
@@ -1,0 +1,3 @@
+package plugins
+
+// empty file w/o build tag so the package "builds"
